Allow for -j > 1 in the build
the number of cores to use may be set in the NCPUS environment
variable. if this is not set, then it attempts to figure out
how many cores are available using nproc. if this command
does not exist, it gives up and defaults to -j1

This can greatly reduce the time to build musenki.